+++ webapp/channels/src/utils/message_html_to_component.test.tsx
@@ -2,6 +2,7 @@
 // See LICENSE.txt for license information.
 
 import {shallow} from 'enzyme';
+import type {AnchorHTMLAttributes} from 'react';
 
 import AtMention from 'components/at_mention';
 import MarkdownImage from 'components/markdown_image';
@@ -9,7 +10,7 @@ import MarkdownImage from 'components/markdown_image';
 import {renderWithContext, screen} from 'tests/react_testing_utils';
 import Constants from 'utils/constants';
 import EmojiMap from 'utils/emoji_map';
-import messageHtmlToComponent from 'utils/message_html_to_component';
+import messageHtmlToComponent, {convertPropsToReactStandard} from 'utils/message_html_to_component';
 import * as TextFormatting from 'utils/text_formatting';
 
 const emptyEmojiMap = new EmojiMap(new Map());
@@ -185,5 +186,60 @@ const myFunction = () => {
 
             expect(container).toHaveTextContent('These are emojis: 🌮 🧑‍🚀');
         });
+    });
+});
+
+describe('convertPropsToReactStandard', () => {
+    test('converts class to className', () => {
+        const inputProps = {class: 'button-class'} as AnchorHTMLAttributes<HTMLAnchorElement>;
+        const expected = {className: 'button-class'};
+
+        expect(convertPropsToReactStandard(inputProps)).toEqual(expected);
+    });
+
+    test('converts for to htmlFor', () => {
+        const inputProps = {for: 'input-id'} as AnchorHTMLAttributes<HTMLAnchorElement>;
+        const expected = {htmlFor: 'input-id'};
+
+        expect(convertPropsToReactStandard(inputProps)).toEqual(expected);
+    });
+
+    test('converts tabindex to tabIndex', () => {
+        const inputProps = {tabindex: '0'} as AnchorHTMLAttributes<HTMLAnchorElement>;
+        const expected = {tabIndex: '0'};
+
+        expect(convertPropsToReactStandard(inputProps)).toEqual(expected);
+    });
+
+    test('converts readonly to readOnly', () => {
+        const inputProps = {readonly: true} as AnchorHTMLAttributes<HTMLAnchorElement>;
+        const expected = {readOnly: true};
+
+        expect(convertPropsToReactStandard(inputProps)).toEqual(expected);
+    });
+
+    test('keeps other properties unchanged', () => {
+        const inputProps = {id: 'unique-id', type: 'text'} as AnchorHTMLAttributes<HTMLAnchorElement>;
+        const expected = {id: 'unique-id', type: 'text'};
+
+        expect(convertPropsToReactStandard(inputProps)).toEqual(expected);
+    });
+
+    test('handles multiple conversions and keeps other properties', () => {
+        const inputProps = {
+            class: 'button-class',
+            for: 'input-id',
+            tabindex: '0',
+            readonly: true,
+            id: 'unique-id',
+        };
+        const expected = {
+            className: 'button-class',
+            htmlFor: 'input-id',
+            tabIndex: '0',
+            readOnly: true,
+            id: 'unique-id',
+        };
+        expect(convertPropsToReactStandard(inputProps)).toEqual(expected);
     });
 });
